Frame specific parts of #19504 (#19512)
